requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Admin==1.5.8
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.25
Werkzeug==2.0.1
WTForms==2.3.3
python-dotenv==0.19.0
pymysql==1.0.2
PyMySQL
cryptography
flask-security-too
Flask-Migrate
bleach

